fix: consolidate to one provider
